+++ packages/orq-rc/RELEASES.md
@@ -7609,3 +7609,13 @@ Based on:
 - [typescript v4.5.0-rc.48] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.48 - packages/orq-rc
+
+## 2026-03-10 17:55:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.750.0 (2.856.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.49] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.49 - packages/orq-rc